refactor(cli): harden openclaw migration tool and tests
