+++ astrbot/cli/utils/__init__.py
@@ -3,16 +3,24 @@ from .basic import (
     check_dashboard,
     get_astrbot_root,
 )
+from .openclaw_migrate import (
+    MemoryEntry,
+    MigrationReport,
+    run_openclaw_migration,
+)
 from .plugin import PluginStatus, build_plug_list, get_git_repo, manage_plugin
 from .version_comparator import VersionComparator
 
 __all__ = [
     "PluginStatus",
     "VersionComparator",
+    "MemoryEntry",
+    "MigrationReport",
     "build_plug_list",
     "check_astrbot_root",
     "check_dashboard",
     "get_astrbot_root",
     "get_git_repo",
     "manage_plugin",
+    "run_openclaw_migration",
 ]
